decompose object & fm
